+++ server/src/test/java/org/opensearch/cluster/node/DiscoveryNodeTests.java
@@ -57,9 +57,8 @@ import java.util.stream.Collectors;
 import static java.util.Collections.emptyMap;
 import static java.util.Collections.emptySet;
 import static org.opensearch.test.NodeRoles.nonRemoteClusterClientNode;
-import static org.opensearch.test.NodeRoles.nonSearchNode;
+import static org.opensearch.test.NodeRoles.nonWarmNode;
 import static org.opensearch.test.NodeRoles.remoteClusterClientNode;
-import static org.opensearch.test.NodeRoles.searchNode;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.greaterThanOrEqualTo;
 import static org.hamcrest.Matchers.hasItem;
@@ -236,12 +235,12 @@ public class DiscoveryNodeTests extends OpenSearchTestCase {
         runTestDiscoveryNodeIsRemoteClusterClient(nonRemoteClusterClientNode(), false);
     }
 
-    public void testDiscoveryNodeIsSearchSet() {
-        runTestDiscoveryNodeIsSearch(searchNode(), true);
+    public void testDiscoveryNodeIsWarmSet() {
+        runTestDiscoveryNodeIsWarm(NodeRoles.warmNode(), true);
     }
 
-    public void testDiscoveryNodeIsSearchUnset() {
-        runTestDiscoveryNodeIsSearch(nonSearchNode(), false);
+    public void testDiscoveryNodeIsWarmUnset() {
+        runTestDiscoveryNodeIsWarm(nonWarmNode(), false);
     }
 
     // Added in 2.0 temporarily, validate the MASTER_ROLE is in the list of known roles.
@@ -263,13 +262,13 @@ public class DiscoveryNodeTests extends OpenSearchTestCase {
         }
     }
 
-    private void runTestDiscoveryNodeIsSearch(final Settings settings, final boolean expected) {
+    private void runTestDiscoveryNodeIsWarm(final Settings settings, final boolean expected) {
         final DiscoveryNode node = DiscoveryNode.createLocal(settings, new TransportAddress(TransportAddress.META_ADDRESS, 9200), "node");
-        assertThat(node.isSearchNode(), equalTo(expected));
+        assertThat(node.isWarmNode(), equalTo(expected));
         if (expected) {
-            assertThat(node.getRoles(), hasItem(DiscoveryNodeRole.SEARCH_ROLE));
+            assertThat(node.getRoles(), hasItem(DiscoveryNodeRole.WARM_ROLE));
         } else {
-            assertThat(node.getRoles(), not(hasItem(DiscoveryNodeRole.SEARCH_ROLE)));
+            assertThat(node.getRoles(), not(hasItem(DiscoveryNodeRole.WARM_ROLE)));
         }
     }
 
@@ -284,9 +283,9 @@ public class DiscoveryNodeTests extends OpenSearchTestCase {
         assertEquals(dynamicRoleName.toLowerCase(Locale.ROOT), dynamicNodeRole.roleNameAbbreviation());
     }
 
-    public void testDiscoveryNodeIsSearchNode() {
-        final Settings settingWithSearchRole = NodeRoles.onlyRole(DiscoveryNodeRole.SEARCH_ROLE);
-        final DiscoveryNode node = DiscoveryNode.createLocal(settingWithSearchRole, buildNewFakeTransportAddress(), "node");
-        assertThat(node.isSearchNode(), equalTo(true));
+    public void testDiscoveryNodeIsWarmNode() {
+        final Settings settingWithWarmRole = NodeRoles.onlyRole(DiscoveryNodeRole.WARM_ROLE);
+        final DiscoveryNode node = DiscoveryNode.createLocal(settingWithWarmRole, buildNewFakeTransportAddress(), "node");
+        assertThat(node.isWarmNode(), equalTo(true));
     }
 }